+++ ui/web/src/pages/channels/channel-instance-form-dialog.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback } from "react";
-import { useForm, Controller } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useTranslation } from "react-i18next";
 import {
@@ -10,26 +10,13 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Switch } from "@/components/ui/switch";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 import type { ChannelInstanceData, ChannelInstanceInput } from "./hooks/use-channel-instances";
 import type { AgentData } from "@/types/agent";
-import { slugify } from "@/lib/slug";
 import { credentialsSchema, configSchema, wizardConfig, type FieldDef } from "./channel-schemas";
-import { ChannelFields } from "./channel-fields";
-import { ChannelScopesInfo } from "./channel-scopes-info";
-import { wizardAuthSteps, wizardConfigSteps, wizardEditConfigs } from "./channel-wizard-registry";
-import { TelegramGroupOverrides } from "./telegram-group-overrides";
+import { wizardAuthSteps, wizardConfigSteps } from "./channel-wizard-registry";
 import { CHANNEL_TYPES } from "@/constants/channels";
 import { channelInstanceSchema, type ChannelInstanceFormData } from "@/schemas/channel.schema";
+import { ChannelInstanceFormStep } from "./channel-instance-form-step";
 
 type WizardStep = "form" | "auth" | "config";
 
@@ -52,7 +39,6 @@ export function ChannelInstanceFormDialog({
 }: ChannelInstanceFormDialogProps) {
   const { t } = useTranslation("channels");
 
-  // Non-form state (dynamic maps + wizard flow)
   const [credsValues, setCredsValues] = useState<Record<string, unknown>>({});
   const [configValues, setConfigValues] = useState<Record<string, unknown>>({});
   const [loading, setLoading] = useState(false);
@@ -78,7 +64,6 @@ export function ChannelInstanceFormDialog({
   const hasWizard = !instance && !!wizard;
   const channelLabel = CHANNEL_TYPES.find((ct) => ct.value === channelType)?.label ?? channelType;
 
-  // Step navigation
   const totalSteps = hasWizard ? 1 + wizard!.steps.length : 1;
   const currentStepNum = step === "form" ? 1 : (wizard?.steps.indexOf(step as "auth" | "config") ?? 0) + 2;
 
@@ -100,7 +85,6 @@ export function ChannelInstanceFormDialog({
       });
       setCredsValues({});
 
-      // Merge schema defaults into config so select fields persist their defaults.
       const ct = instance?.channel_type ?? "telegram";
       const schema = configSchema[ct] ?? [];
       const defaults: Record<string, unknown> = {};
@@ -108,9 +92,8 @@ export function ChannelInstanceFormDialog({
         if (f.defaultValue !== undefined) defaults[f.key] = f.defaultValue;
       }
       const merged: Record<string, unknown> = { ...defaults, ...(instance?.config ?? {}) };
-      // Convert boolean values to strings for select fields that use "true"/"false" options
       const boolSelectKeys = new Set(
-        schema.filter((f) => f.type === "select" && f.options?.some((o) => o.value === "true")).map((f) => f.key),
+        schema.filter((f: FieldDef) => f.type === "select" && f.options?.some((o) => o.value === "true")).map((f: FieldDef) => f.key),
       );
       for (const key of boolSelectKeys) {
         if (typeof merged[key] === "boolean") merged[key] = String(merged[key]);
@@ -124,7 +107,6 @@ export function ChannelInstanceFormDialog({
     }
   }, [open, instance, agents, form]);
 
-  // Auto-advance from auth to next step on completion
   useEffect(() => {
     if (step !== "auth" || !authCompleted) return;
     const next = getNextWizardStep("auth");
@@ -143,8 +125,6 @@ export function ChannelInstanceFormDialog({
     setConfigValues((prev) => ({ ...prev, [key]: value }));
   }, []);
 
-  // Convert select fields with "true"/"false"/"inherit" values to proper JSON types.
-  // "inherit" → remove key (nil on Go side), "true"/"false" → boolean.
   const coerceBoolSelects = (cfg: Record<string, unknown>, schema: FieldDef[]) => {
     const boolSelectKeys = new Set(
       schema.filter((f) => f.type === "select" && f.options?.some((o) => o.value === "true")).map((f) => f.key),
@@ -153,16 +133,16 @@ export function ChannelInstanceFormDialog({
       const v = cfg[key];
       if (v === "true") cfg[key] = true;
       else if (v === "false") cfg[key] = false;
-      else delete cfg[key]; // "inherit" or unset
+      else delete cfg[key];
     }
   };
 
   const handleSubmit = form.handleSubmit(async (values) => {
     if (!instance) {
       const schema = credentialsSchema[values.channelType] ?? [];
-      const missing = schema.filter((f) => f.required && !credsValues[f.key]);
+      const missing = schema.filter((f: FieldDef) => f.required && !credsValues[f.key]);
       if (missing.length > 0) {
-        setError(t("form.errors.requiredFields", { fields: missing.map((f) => f.label).join(", ") }));
+        setError(t("form.errors.requiredFields", { fields: missing.map((f: FieldDef) => f.label).join(", ") }));
         return;
       }
     }
@@ -234,15 +214,14 @@ export function ChannelInstanceFormDialog({
   };
 
   const canClose = step !== "auth";
-  const credsFields = credentialsSchema[channelType] ?? [];
-  const excludeSet = new Set(wizard?.excludeConfigFields ?? []);
-  const cfgFields = configSchema[channelType] ?? [];
-  const formCfgFields = excludeSet.size > 0 ? cfgFields.filter((f) => !excludeSet.has(f.key)) : cfgFields;
-
-  // Lookup registered step components for current channel type
   const AuthStep = wizardAuthSteps[channelType];
   const ConfigStep = wizardConfigSteps[channelType];
-  const EditConfig = wizardEditConfigs[channelType];
+
+  const submitLabel = loading
+    ? t("form.saving")
+    : instance
+      ? t("form.update")
+      : (wizard?.createLabel ? t(wizard.createLabel) : t("form.create"));
 
   const dialogTitle = instance
     ? t("form.editTitle")
@@ -251,8 +230,6 @@ export function ChannelInstanceFormDialog({
       : step === "auth"
         ? t("form.authenticate", { label: channelLabel })
         : t("form.configure", { label: channelLabel });
-
-  const { register, control, formState: { errors } } = form;
 
   return (
     <Dialog open={open} onOpenChange={(v) => { if (!loading && canClose) onOpenChange(v); }}>
@@ -266,153 +243,24 @@ export function ChannelInstanceFormDialog({
           )}
         </DialogHeader>
 
-        {/* === FORM STEP === */}
         {step === "form" && (
-          <>
-            <div className="grid gap-4 py-2 -mx-4 px-4 sm:-mx-6 sm:px-6 overflow-y-auto min-h-0">
-              <div className="grid gap-1.5">
-                <Label htmlFor="ci-name">{t("form.key")}</Label>
-                <Input
-                  id="ci-name"
-                  {...register("name", {
-                    setValueAs: (v: string) => slugify(v),
-                  })}
-                  onChange={(e) => form.setValue("name", slugify(e.target.value), { shouldValidate: true })}
-                  value={form.watch("name")}
-                  placeholder={t("form.keyPlaceholder")}
-                  disabled={!!instance}
-                />
-                {errors.name && <p className="text-xs text-destructive">{errors.name.message}</p>}
-                <p className="text-xs text-muted-foreground">{t("form.keyHint")}</p>
-              </div>
-
-              <div className="grid gap-1.5">
-                <Label htmlFor="ci-display">{t("form.displayName")}</Label>
-                <Input
-                  id="ci-display"
-                  {...register("displayName")}
-                  placeholder={t("form.displayNamePlaceholder")}
-                />
-              </div>
-
-              <div className="grid gap-1.5">
-                <Label>{t("form.channelType")}</Label>
-                <Controller
-                  control={control}
-                  name="channelType"
-                  render={({ field }) => (
-                    <Select value={field.value} onValueChange={field.onChange} disabled={!!instance}>
-                      <SelectTrigger><SelectValue /></SelectTrigger>
-                      <SelectContent>
-                        {CHANNEL_TYPES.map((ct) => (
-                          <SelectItem key={ct.value} value={ct.value}>{ct.label}</SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                  )}
-                />
-              </div>
-
-              <div className="grid gap-1.5">
-                <Label>{t("form.agent")}</Label>
-                <Controller
-                  control={control}
-                  name="agentId"
-                  render={({ field }) => (
-                    <Select value={field.value} onValueChange={field.onChange}>
-                      <SelectTrigger><SelectValue placeholder={t("form.selectAgent")} /></SelectTrigger>
-                      <SelectContent>
-                        {agents.map((a) => (
-                          <SelectItem key={a.id} value={a.id}>{a.display_name || a.agent_key}</SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                  )}
-                />
-                {errors.agentId && <p className="text-xs text-destructive">{errors.agentId.message}</p>}
-              </div>
-
-              {credsFields.length > 0 && (
-                <fieldset className="rounded-md border p-3 space-y-3">
-                  <legend className="px-1 text-sm font-medium">
-                    {t("form.credentials")}
-                    {instance && <span className="text-xs font-normal text-muted-foreground ml-1">{t("form.credentialsHint")}</span>}
-                  </legend>
-                  <ChannelFields fields={credsFields} values={credsValues} onChange={handleCredsChange} idPrefix="ci-cred" isEdit={!!instance} contextValues={configValues} />
-                  <p className="text-xs text-muted-foreground">{t("form.credentialsEncrypted")}</p>
-                </fieldset>
-              )}
-
-              <ChannelScopesInfo channelType={channelType} />
-
-              {/* Auth status indicator (edit mode, channels with auth wizard step) */}
-              {instance && wizard?.steps.includes("auth") && (
-                <div className="rounded-md border border-blue-200 bg-blue-50 dark:border-blue-900 dark:bg-blue-950 p-3">
-                  <div className="flex items-center gap-2">
-                    <span className={`h-2 w-2 rounded-full ${instance.has_credentials ? "bg-green-500" : "bg-amber-500"}`} />
-                    <span className="text-sm">
-                      {instance.has_credentials
-                        ? t("form.authStatus.authenticated")
-                        : t("form.authStatus.notAuthenticated")}
-                    </span>
-                    {!instance.has_credentials && (
-                      <span className="text-xs text-muted-foreground ml-1">{t("form.authStatus.useQrHint")}</span>
-                    )}
-                  </div>
-                </div>
-              )}
-
-              {/* Wizard info banner (create mode) */}
-              {hasWizard && wizard?.formBanner && (
-                <div className="rounded-md border border-blue-200 bg-blue-50 dark:border-blue-900 dark:bg-blue-950 p-3">
-                  <p className="text-sm text-muted-foreground">{t(wizard.formBanner)}</p>
-                </div>
-              )}
-
-              {formCfgFields.length > 0 && (
-                <fieldset className="rounded-md border p-3 space-y-3">
-                  <legend className="px-1 text-sm font-medium">{t("form.configuration")}</legend>
-                  <ChannelFields fields={formCfgFields} values={configValues} onChange={handleConfigChange} idPrefix="ci-cfg" />
-                  {instance && EditConfig && <EditConfig instance={instance} configValues={configValues} onConfigChange={handleConfigChange} />}
-                </fieldset>
-              )}
-
-              {/* Telegram group/topic overrides */}
-              {channelType === "telegram" && (
-                <TelegramGroupOverrides
-                  groups={(configValues.groups as Record<string, Record<string, unknown>>) ?? {}}
-                  onChange={(groups) => {
-                    setConfigValues((prev) => ({
-                      ...prev,
-                      groups: Object.keys(groups).length > 0 ? groups : undefined,
-                    }));
-                  }}
-                />
-              )}
-
-              <div className="flex items-center gap-2">
-                <Controller
-                  control={control}
-                  name="enabled"
-                  render={({ field }) => (
-                    <Switch id="ci-enabled" checked={field.value} onCheckedChange={field.onChange} />
-                  )}
-                />
-                <Label htmlFor="ci-enabled">{t("form.enabled")}</Label>
-              </div>
-              {error && <p className="text-sm text-destructive">{error}</p>}
-            </div>
-
-            <DialogFooter>
-              <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>{t("form.cancel")}</Button>
-              <Button onClick={handleSubmit} disabled={loading}>
-                {loading ? t("form.saving") : instance ? t("form.update") : (wizard?.createLabel ? t(wizard.createLabel) : t("form.create"))}
-              </Button>
-            </DialogFooter>
-          </>
+          <ChannelInstanceFormStep
+            form={form}
+            instance={instance}
+            agents={agents}
+            credsValues={credsValues}
+            configValues={configValues}
+            onCredsChange={handleCredsChange}
+            onConfigChange={handleConfigChange}
+            setConfigValues={setConfigValues}
+            error={error}
+            loading={loading}
+            onCancel={() => onOpenChange(false)}
+            onSubmit={handleSubmit}
+            submitLabel={submitLabel}
+          />
         )}
 
-        {/* === AUTH STEP (rendered by registered component) === */}
         {step === "auth" && createdInstanceId && AuthStep && (
           <AuthStep
             instanceId={createdInstanceId}
@@ -421,7 +269,6 @@ export function ChannelInstanceFormDialog({
           />
         )}
 
-        {/* === CONFIG STEP (rendered by registered component) === */}
         {step === "config" && createdInstanceId && ConfigStep && (
           <>
             <div className="py-2 -mx-4 px-4 sm:-mx-6 sm:px-6 overflow-y-auto min-h-0">
