+++ ui/web/src/pages/cli-credentials/cli-credential-form-dialog.tsx
@@ -1,26 +1,23 @@
 import type { ManualEnvEntry } from "./cli-credential-env-vars-section";
 import { useState, useEffect } from "react";
-import { useForm, Controller } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useTranslation } from "react-i18next";
 import {
   Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
 import { Label } from "@/components/ui/label";
-import { Switch } from "@/components/ui/switch";
 import {
   Select, SelectContent, SelectItem, SelectTrigger, SelectValue,
 } from "@/components/ui/select";
-import { Search, Check, AlertCircle } from "lucide-react";
 import { useHttp } from "@/hooks/use-ws";
 import { useAgents } from "@/pages/agents/hooks/use-agents";
 import type { SecureCLIBinary, CLICredentialInput, CLIPreset } from "./hooks/use-cli-credentials";
 import { CliCredentialEnvVarsSection } from "./cli-credential-env-vars-section";
+import { CliCredentialBinaryFields } from "./cli-credential-binary-fields";
+import { CliCredentialScopeFields } from "./cli-credential-scope-fields";
 import { cliCredentialSchema, type CliCredentialFormData } from "@/schemas/credential.schema";
-
 
 interface Props {
   open: boolean;
@@ -31,7 +28,6 @@ interface Props {
 }
 
 const NONE_PRESET = "__none__";
-const GLOBAL_AGENT = "__global__";
 const ENV_KEY_PATTERN = /^[A-Za-z_][A-Za-z0-9_]*$/;
 
 export function CliCredentialFormDialog({ open, onOpenChange, credential, presets, onSubmit }: Props) {
@@ -40,7 +36,6 @@ export function CliCredentialFormDialog({ open, onOpenChange, credential, preset
   const http = useHttp();
   const { agents } = useAgents();
 
-  // Non-form state: preset selection, env vars, binary check, loading
   const [selectedPreset, setSelectedPreset] = useState(NONE_PRESET);
   const [envValues, setEnvValues] = useState<Record<string, string>>({});
   const [manualEnvEntries, setManualEnvEntries] = useState<ManualEnvEntry[]>([]);
@@ -73,8 +68,6 @@ export function CliCredentialFormDialog({ open, onOpenChange, credential, preset
     },
   });
 
-  const { register, control, formState: { errors }, setValue, watch } = form;
-  const binaryName = watch("binaryName");
 
   useEffect(() => {
     if (!open) return;
@@ -144,7 +137,7 @@ export function CliCredentialFormDialog({ open, onOpenChange, credential, preset
   };
 
   const handleCheckBinary = async () => {
-    const name = binaryName.trim();
+    const name = form.getValues("binaryName").trim();
     if (!name) return;
     setChecking(true);
     setCheckResult(null);
@@ -154,7 +147,7 @@ export function CliCredentialFormDialog({ open, onOpenChange, credential, preset
         { binary_name: name },
       );
       setCheckResult(res);
-      if (res.found && res.path) setValue("binaryPath", res.path);
+      if (res.found && res.path) form.setValue("binaryPath", res.path);
     } catch {
       setCheckResult({ found: false, error: t("form.binaryNotFound") });
     } finally {
@@ -219,7 +212,6 @@ export function CliCredentialFormDialog({ open, onOpenChange, credential, preset
         </DialogHeader>
 
         <div className="grid gap-4 py-2 -mx-4 px-4 sm:-mx-6 sm:px-6 overflow-y-auto min-h-0">
-          {/* Preset selector — only on create */}
           {!isEdit && presetEntries.length > 0 && (
             <div className="grid gap-1.5">
               <Label>{t("form.preset")}</Label>
@@ -255,160 +247,28 @@ export function CliCredentialFormDialog({ open, onOpenChange, credential, preset
             setManualEnvEntries={setManualEnvEntries}
           />
 
-          {/* Binary name + check button */}
-          <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-            <div className="grid gap-1.5">
-              <Label htmlFor="cc-name">{t("form.binaryName")}</Label>
-              <div className="flex gap-1.5">
-                <Input
-                  id="cc-name"
-                  {...register("binaryName", {
-                    onChange: () => setCheckResult(null),
-                  })}
-                  placeholder={t("placeholders.binaryName")}
-                  className="text-base md:text-sm"
-                />
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="icon"
-                  className="shrink-0"
-                  disabled={!binaryName.trim() || checking}
-                  onClick={handleCheckBinary}
-                  title={t("form.checkBinary")}
-                >
-                  <Search className="h-4 w-4" />
-                </Button>
-              </div>
-              {errors.binaryName && <p className="text-xs text-destructive">{errors.binaryName.message}</p>}
-              {checkResult && (
-                <p className={`text-xs flex items-center gap-1 ${checkResult.found ? "text-green-600 dark:text-green-400" : "text-destructive"}`}>
-                  {checkResult.found ? <Check className="h-3 w-3" /> : <AlertCircle className="h-3 w-3" />}
-                  {checkResult.found
-                    ? t("form.binaryFound", { path: checkResult.path })
-                    : (checkResult.error || t("form.binaryNotFound"))}
-                </p>
-              )}
-              {checking && <p className="text-xs text-muted-foreground">{t("form.checking")}</p>}
-            </div>
-            <div className="grid gap-1.5">
-              <Label htmlFor="cc-path">
-                {t("form.binaryPath")} <span className="text-xs text-muted-foreground">({tc("optional")})</span>
-              </Label>
-              <Input
-                id="cc-path"
-                {...register("binaryPath")}
-                placeholder={t("placeholders.binaryPath")}
-                className="text-base md:text-sm"
-              />
-              <p className="text-xs text-muted-foreground">{t("form.binaryPathHint")}</p>
-            </div>
-          </div>
+          <CliCredentialBinaryFields
+            form={form}
+            checking={checking}
+            checkResult={checkResult}
+            onCheckBinary={handleCheckBinary}
+          />
 
-          <div className="grid gap-1.5">
-            <Label htmlFor="cc-desc">{tc("description")}</Label>
-            <Textarea
-              id="cc-desc"
-              {...register("description")}
-              placeholder={t("placeholders.description")}
-              rows={2}
-              className="text-base md:text-sm"
-            />
-          </div>
-
-          <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-            <div className="grid gap-1.5">
-              <Label htmlFor="cc-deny-args">
-                {t("form.denyArgs")} <span className="text-xs text-muted-foreground">({t("form.commaSeparated")})</span>
-              </Label>
-              <Input
-                id="cc-deny-args"
-                {...register("denyArgs")}
-                placeholder={t("placeholders.denyArgs")}
-                className="text-base md:text-sm"
-              />
-            </div>
-            <div className="grid gap-1.5">
-              <Label htmlFor="cc-timeout">{t("form.timeout")}</Label>
-              <Input
-                id="cc-timeout"
-                type="number"
-                min={1}
-                {...register("timeout", { valueAsNumber: true })}
-                className="text-base md:text-sm"
-              />
-            </div>
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label htmlFor="cc-deny-verbose">
-              {t("form.denyVerbose")} <span className="text-xs text-muted-foreground">({t("form.commaSeparated")})</span>
-            </Label>
-            <Input
-              id="cc-deny-verbose"
-              {...register("denyVerbose")}
-              placeholder={t("placeholders.denyVerbose")}
-              className="text-base md:text-sm"
-            />
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label htmlFor="cc-tips">{t("form.tips")}</Label>
-            <Textarea
-              id="cc-tips"
-              {...register("tips")}
-              placeholder={t("placeholders.tips")}
-              rows={2}
-              className="text-base md:text-sm"
-            />
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label>
-              {t("form.agentId")} <span className="text-xs text-muted-foreground">({t("form.agentIdHint")})</span>
-            </Label>
-            <Controller
-              control={control}
-              name="agentId"
-              render={({ field }) => (
-                <Select
-                  value={field.value || GLOBAL_AGENT}
-                  onValueChange={(v) => field.onChange(v === GLOBAL_AGENT ? "" : v)}
-                >
-                  <SelectTrigger className="text-base md:text-sm">
-                    <SelectValue placeholder={t("placeholders.agentId")} />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value={GLOBAL_AGENT}>{t("placeholders.agentId")}</SelectItem>
-                    {agents.map((a) => (
-                      <SelectItem key={a.id} value={a.id}>
-                        {a.display_name || a.agent_key || a.id}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              )}
-            />
-          </div>
-
-          <div className="flex items-center gap-2">
-            <Controller
-              control={control}
-              name="enabled"
-              render={({ field }) => (
-                <Switch id="cc-enabled" checked={field.value} onCheckedChange={field.onChange} />
-              )}
-            />
-            <Label htmlFor="cc-enabled">{tc("enabled")}</Label>
-          </div>
+          <CliCredentialScopeFields form={form} agents={agents} />
 
           {error && <p className="text-sm text-destructive">{error}</p>}
         </div>
 
         <DialogFooter>
-          <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>{tc("cancel")}</Button>
+          <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>
+            {tc("cancel")}
+          </Button>
           <Button onClick={handleSubmit} disabled={loading}>
-            {loading ? tc("saving") : isEdit ? tc("update") : tc("create")}
+            {loading
+              ? tc("saving")
+              : isEdit
+                ? tc("update")
+                : tc("create")}
           </Button>
         </DialogFooter>
       </DialogContent>
