+++ ui/web/src/pages/mcp/mcp-form-dialog.tsx
@@ -11,30 +11,11 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Switch } from "@/components/ui/switch";
-import { KeyValueEditor } from "@/components/shared/key-value-editor";
 import type { MCPServerData, MCPServerInput } from "./hooks/use-mcp";
-import { slugify, isValidSlug } from "@/lib/slug";
+import { isValidSlug } from "@/lib/slug";
 import { mcpFormSchema, type MCPFormData } from "@/schemas/mcp.schema";
-
-/** Header keys whose values should be masked in the form. */
-const SENSITIVE_HEADER_RE = /^(authorization|x-api-key|api-key|bearer|token|secret|password|credential)/i;
-
-/** Env var keys whose values should be masked in the form. */
-const SENSITIVE_ENV_RE = /^.*(key|secret|token|password|credential).*$/i;
-
-const isSensitiveHeader = (key: string) => SENSITIVE_HEADER_RE.test(key.trim());
-const isSensitiveEnv = (key: string) => SENSITIVE_ENV_RE.test(key.trim());
-
-interface MCPFormDialogProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  server?: MCPServerData | null;
-  onSubmit: (data: MCPServerInput) => Promise<unknown>;
-  onTest: (data: { transport: string; command?: string; args?: string[]; url?: string; headers?: Record<string, string>; env?: Record<string, string> }) => Promise<{ success: boolean; tool_count?: number; error?: string }>;
-}
+import { McpConnectionFields } from "./mcp-connection-fields";
+import { McpSettingsFields } from "./mcp-settings-fields";
 
 /** Split a string into shell-like tokens, treating commas and spaces outside quotes as delimiters. */
 function splitShellTokens(input: string): string[] {
@@ -47,11 +28,20 @@ function splitShellTokens(input: string): string[] {
   return tokens.filter(Boolean);
 }
 
-const TRANSPORTS = [
-  { value: "stdio", label: "stdio" },
-  { value: "sse", label: "SSE" },
-  { value: "streamable-http", label: "Streamable HTTP" },
-] as const;
+interface MCPFormDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  server?: MCPServerData | null;
+  onSubmit: (data: MCPServerInput) => Promise<unknown>;
+  onTest: (data: {
+    transport: string;
+    command?: string;
+    args?: string[];
+    url?: string;
+    headers?: Record<string, string>;
+    env?: Record<string, string>;
+  }) => Promise<{ success: boolean; tool_count?: number; error?: string }>;
+}
 
 export function MCPFormDialog({ open, onOpenChange, server, onSubmit, onTest }: MCPFormDialogProps) {
   const { t } = useTranslation("mcp");
@@ -79,19 +69,13 @@ export function MCPFormDialog({ open, onOpenChange, server, onSubmit, onTest }: 
     },
   });
 
-  const { register, watch, setValue, reset, handleSubmit: rhfHandleSubmit } = form;
+  const { watch, reset, handleSubmit: rhfHandleSubmit } = form;
   const transport = watch("transport");
-  const name = watch("name");
   const command = watch("command");
   const args = watch("args");
   const url = watch("url");
   const headers = watch("headers") as Record<string, string>;
   const env = watch("env") as Record<string, string>;
-  const enabled = watch("enabled");
-  const requireUserCreds = watch("requireUserCreds");
-  const timeout = watch("timeout");
-  const toolPrefix = watch("toolPrefix");
-
   const isStdio = transport === "stdio";
 
   useEffect(() => {
@@ -131,32 +115,19 @@ export function MCPFormDialog({ open, onOpenChange, server, onSubmit, onTest }: 
       }
     }
 
-    const parsedHeaders = !isStdio && Object.keys(headers).length > 0
-      ? (headers as Record<string, string>)
-      : undefined;
-    const parsedEnv = Object.keys(env).length > 0
-      ? (env as Record<string, string>)
-      : undefined;
     return {
       transport,
       command: isStdio ? resolvedCommand : undefined,
       args: parsedArgs,
       url: !isStdio ? url.trim() : undefined,
-      headers: parsedHeaders,
-      env: parsedEnv,
+      headers: !isStdio && Object.keys(headers).length > 0 ? headers : undefined,
+      env: Object.keys(env).length > 0 ? env : undefined,
     };
   };
 
   const handleTest = async () => {
-    if (isStdio && !command.trim()) {
-      setError(t("form.errors.commandRequired"));
-      return;
-    }
-    if (!isStdio && !url.trim()) {
-      setError(t("form.errors.urlRequired"));
-      return;
-    }
-
+    if (isStdio && !command.trim()) { setError(t("form.errors.commandRequired")); return; }
+    if (!isStdio && !url.trim()) { setError(t("form.errors.urlRequired")); return; }
     setTesting(true);
     setError("");
     setTestResult(null);
@@ -171,27 +142,17 @@ export function MCPFormDialog({ open, onOpenChange, server, onSubmit, onTest }: 
   };
 
   const handleSubmit = rhfHandleSubmit(async (data) => {
-    if (!isValidSlug(data.name.trim())) {
-      setError(t("form.errors.nameSlug"));
-      return;
-    }
-    if (isStdio && !data.command.trim()) {
-      setError(t("form.errors.commandRequired"));
-      return;
-    }
-    if (!isStdio && !data.url.trim()) {
-      setError(t("form.errors.urlRequired"));
-      return;
-    }
+    if (!isValidSlug(data.name.trim())) { setError(t("form.errors.nameSlug")); return; }
+    if (isStdio && !data.command.trim()) { setError(t("form.errors.commandRequired")); return; }
+    if (!isStdio && !data.url.trim()) { setError(t("form.errors.urlRequired")); return; }
 
     setLoading(true);
     setError("");
     try {
-      const conn = buildConnectionData();
       await onSubmit({
         name: data.name.trim(),
         display_name: data.displayName.trim() || undefined,
-        ...conn,
+        ...buildConnectionData(),
         tool_prefix: data.toolPrefix.trim() || undefined,
         timeout_sec: data.timeout,
         settings: { require_user_credentials: data.requireUserCreds },
@@ -213,141 +174,30 @@ export function MCPFormDialog({ open, onOpenChange, server, onSubmit, onTest }: 
         </DialogHeader>
 
         <div className="grid gap-4 py-2 -mx-4 px-4 sm:-mx-6 sm:px-6 overflow-y-auto min-h-0">
-          <div className="grid gap-1.5">
-            <Label htmlFor="mcp-name">{t("form.name")}</Label>
-            <Input
-              id="mcp-name"
-              value={name}
-              onChange={(e) => setValue("name", slugify(e.target.value))}
-              placeholder="my-mcp-server"
-            />
-            <p className="text-xs text-muted-foreground">{t("form.nameHint")}</p>
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label htmlFor="mcp-display">{t("form.displayName")}</Label>
-            <Input id="mcp-display" placeholder={t("form.displayNamePlaceholder")} {...register("displayName")} />
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label>{t("form.transport")}</Label>
-            <div className="flex gap-2">
-              {TRANSPORTS.map((tr) => (
-                <Button
-                  key={tr.value}
-                  type="button"
-                  variant={transport === tr.value ? "default" : "outline"}
-                  size="sm"
-                  onClick={() => setValue("transport", tr.value)}
-                >
-                  {tr.label}
-                </Button>
-              ))}
-            </div>
-          </div>
-
-          {isStdio ? (
-            <>
-              <div className="grid gap-1.5">
-                <Label htmlFor="mcp-cmd">{t("form.command")}</Label>
-                <Input id="mcp-cmd" placeholder="npx" className="font-mono" {...register("command")} />
-              </div>
-              <div className="grid gap-1.5">
-                <Label htmlFor="mcp-args">{t("form.args")}</Label>
-                <Input id="mcp-args" placeholder={t("form.argsPlaceholder")} className="font-mono" {...register("args")} />
-              </div>
-            </>
-          ) : (
-            <>
-              <div className="grid gap-1.5">
-                <Label htmlFor="mcp-url">{t("form.url")}</Label>
-                <Input id="mcp-url" placeholder="http://localhost:3001/sse" className="font-mono" {...register("url")} />
-              </div>
-              <div className="grid gap-1.5">
-                <Label>{t("form.headers")}</Label>
-                <KeyValueEditor
-                  value={headers}
-                  onChange={(v) => setValue("headers", v)}
-                  keyPlaceholder={t("form.headerKeyPlaceholder")}
-                  valuePlaceholder={t("form.headerValuePlaceholder")}
-                  addLabel={t("form.addHeader")}
-                  maskValue={isSensitiveHeader}
-                />
-              </div>
-            </>
-          )}
-
-          <div className="grid gap-1.5">
-            <Label>{t("form.env")}</Label>
-            <KeyValueEditor
-              value={env}
-              onChange={(v) => setValue("env", v)}
-              keyPlaceholder={t("form.envKeyPlaceholder")}
-              valuePlaceholder={t("form.envValuePlaceholder")}
-              addLabel={t("form.addVariable")}
-              maskValue={isSensitiveEnv}
-            />
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label htmlFor="mcp-prefix">{t("form.toolPrefix")}</Label>
-            <div className="flex">
-              <span className="inline-flex items-center px-2.5 rounded-l-md border border-r-0 border-input bg-muted text-muted-foreground text-sm font-mono">mcp_</span>
-              <Input
-                id="mcp-prefix"
-                value={toolPrefix}
-                onChange={(e) => setValue("toolPrefix", e.target.value.replace(/[^a-z0-9_]/g, ""))}
-                placeholder={name.replace(/-/g, "_") || "auto"}
-                className="rounded-l-none font-mono"
-              />
-            </div>
-            <p className="text-xs text-muted-foreground">{t("form.toolPrefixHint")} Tools: <code className="text-[10px]">mcp_&#123;prefix&#125;__&#123;tool&#125;</code></p>
-          </div>
-
-          <div className="grid gap-1.5">
-            <Label htmlFor="mcp-timeout">{t("form.timeout")}</Label>
-            <Input
-              id="mcp-timeout"
-              type="number"
-              value={timeout}
-              onChange={(e) => setValue("timeout", Number(e.target.value))}
-              min={1}
-            />
-          </div>
-
-          <div className="flex items-center gap-2">
-            <Switch id="mcp-enabled" checked={enabled} onCheckedChange={(v) => setValue("enabled", v)} />
-            <Label htmlFor="mcp-enabled">{t("form.enabled")}</Label>
-          </div>
-
-          <div className="space-y-1">
-            <div className="flex items-center gap-2">
-              <Switch id="mcp-require-creds" checked={requireUserCreds} onCheckedChange={(v) => setValue("requireUserCreds", v)} />
-              <Label htmlFor="mcp-require-creds">{t("form.requireUserCredentials")}</Label>
-            </div>
-            <p className="text-xs text-muted-foreground pl-9">{t("form.requireUserCredentialsHint")}</p>
-          </div>
-
+          <McpConnectionFields form={form} />
+          <McpSettingsFields form={form} />
           {error && <p className="text-sm text-destructive">{error}</p>}
         </div>
 
         <DialogFooter className="flex-col sm:flex-row gap-2">
           <div className="flex items-center gap-2 mr-auto">
             <Button type="button" variant="secondary" size="sm" onClick={handleTest} disabled={loading || testing}>
-              {testing ? <><Loader2 className="h-3.5 w-3.5 animate-spin mr-1" /> {t("form.testing")}</> : t("form.testConnection")}
+              {testing
+                ? <><Loader2 className="h-3.5 w-3.5 animate-spin mr-1" />{t("form.testing")}</>
+                : t("form.testConnection")}
             </Button>
             {testResult && (
               <span className={`flex items-center gap-1 text-xs ${testResult.success ? "text-emerald-600 dark:text-emerald-400" : "text-destructive"}`}>
-                {testResult.success ? (
-                  <><CheckCircle2 className="h-3.5 w-3.5" /> {t("form.toolsFound", { count: testResult.tool_count })}</>
-                ) : (
-                  <><XCircle className="h-3.5 w-3.5" /> {testResult.error}</>
-                )}
+                {testResult.success
+                  ? <><CheckCircle2 className="h-3.5 w-3.5" />{t("form.toolsFound", { count: testResult.tool_count })}</>
+                  : <><XCircle className="h-3.5 w-3.5" />{testResult.error}</>}
               </span>
             )}
           </div>
           <div className="flex gap-2">
-            <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>{t("form.cancel")}</Button>
+            <Button variant="outline" onClick={() => onOpenChange(false)} disabled={loading}>
+              {t("form.cancel")}
+            </Button>
             <Button onClick={handleSubmit} disabled={loading}>
               {loading ? t("form.saving") : server ? t("form.update") : t("form.create")}
             </Button>
